requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.6.0
hiredis==2.0.0
huggingface-hub==0.5.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
joblib==1.1.0
kombu==5.2.4
msgpack==1.0.3
nltk==3.7
numpy==1.22.3
packaging==21.3
phonenumbers==8.12.41
Pillow==9.1.0
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytz==2021.3
PyYAML==6.0
redis==4.2.2
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zope.interface==5.4.0
